Implemented BookController,Book validattion

+++ online-bookstore/src/main/java/com/onlineBookStore/dto/BookDTO.java
@@ -1,5 +1,9 @@
 package com.onlineBookStore.dto;
 
+
+
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -8,7 +12,7 @@ public class BookDTO {
 	private Long id;
 	private String title;
 	private String author;
-	private Double price;
-	private Integer stock;
+	private BigDecimal price;
+	private BigDecimal stock;
 	
 }
